Move token range tests.
